requirements.txt
fastapi
pyautogen
pydantic
websocket
uvicorn
jupyter_client
jupyter_kernel_gateway
ipykernel